Add fast filesystem for docker
This lets you set a single device file which, if present, will be used to
back /var/lib/docker.

The variable should probably be expanded and checked to see if it refers
to multiple devices; under those circmstances we should add all devices
to a VG and build the FS on that.

It'd be okay to default this to /dev/nvme0n1 since it's a no-op on [VB]M.Standard*
systems.